refactor: 엔티티 네이밍 수정 및 엔티티 위치 이동
