+++ src/main/java/com/example/daobe/objet/domain/repository/ObjetSharerRepository.java
@@ -1,0 +1,7 @@
+package com.example.daobe.objet.domain.repository;
+
+import com.example.daobe.objet.domain.ObjetSharer;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ObjetSharerRepository extends JpaRepository<ObjetSharer, Long> {
+}
