--- src/main/java/com/example/daobe/shared/repository/UserObjetRepository.java
@@ -1,7 +1,0 @@
-package com.example.daobe.shared.repository;
-
-import com.example.daobe.shared.entity.UserObjet;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserObjetRepository extends JpaRepository<UserObjet, Long> {
-}
